✨ Compute score in rolling manner
instead of resetting at midnight.
It will allow to have a meaningful value at any point of the day.

In simple cases (like offpeak hours during the night), it will
especially enable to have a value that shows reality instead of an
always degrading value as soon as we consume electricity during peak
time.

Fix #57

Change-Id: I7100cf7cca783a94ae67df32ba6516e18246563b